# spring-java-backend/api/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/users/me":{"get":{"tags":["user-controller"],"operationId":"getUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetUserDTO"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutUserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetUserDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUsers","responses":{"200":{"description":"OK"}}}},"/api/v1/users/archived":{"put":{"tags":["user-controller"],"operationId":"activateUsers","parameters":[{"name":"ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/gradplans/archived":{"get":{"tags":["grad-plan-controller"],"operationId":"getArchivedPlans","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetGradPlanDTO"}}}}}}},"put":{"tags":["grad-plan-controller"],"operationId":"activatePlans","parameters":[{"name":"ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/classplans/archived":{"get":{"tags":["class-plan-controller"],"operationId":"getArchivedPlans_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetClassPlanDTO"}}}}}}},"put":{"tags":["class-plan-controller"],"operationId":"activatePlans_1","parameters":[{"name":"ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users/register":{"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostUserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetUserDTO"}}}}}}},"/api/v1/users/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO.Login"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/gradplans":{"get":{"tags":["grad-plan-controller"],"operationId":"getPlans","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetGradPlanDTO"}}}}}}},"post":{"tags":["grad-plan-controller"],"operationId":"createPlan","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file-name":{"type":"string"},"file":{"type":"string","format":"binary"}},"required":["file","file-name"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetGradPlanDTO"}}}}}},"delete":{"tags":["grad-plan-controller"],"operationId":"deletePlans","parameters":[{"name":"ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/classplans":{"get":{"tags":["class-plan-controller"],"operationId":"getPlans_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetClassPlanDTO"}}}}}}},"post":{"tags":["class-plan-controller"],"operationId":"createPlan_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostClassPlanDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetClassPlanDTO"}}}}}},"delete":{"tags":["class-plan-controller"],"operationId":"deletePlans_1","parameters":[{"name":"ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/gradplans/{id}":{"get":{"tags":["grad-plan-controller"],"operationId":"getPlan","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetGradPlanDTO"}}}}}}},"/api/v1/classplans/{id}":{"get":{"tags":["class-plan-controller"],"operationId":"getPlan_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetClassPlanDTO"}}}}}}}},"components":{"schemas":{"PutUserDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"GetUserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"PostUserDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"UserDTO.Login":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"archived":{"type":"boolean"}}},"GetGradPlanDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fileName":{"type":"string"},"programName":{"type":"string"},"majorName":{"type":"string"},"creditsCompleted":{"type":"integer","format":"int64"},"creditsRequired":{"type":"integer","format":"int64"},"cgpa":{"type":"number","format":"double"},"programLevel":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"PostAvailabilityDTO":{"type":"object","properties":{"days":{"type":"array","items":{"$ref":"#/components/schemas/PostAvailabilityDayDTO"}}}},"PostAvailabilityDayDTO":{"type":"object","properties":{"day":{"type":"string"},"hours":{"type":"array","items":{"$ref":"#/components/schemas/PostAvailabilityHourDTO"}}}},"PostAvailabilityHourDTO":{"type":"object","properties":{"hourOfTheDay":{"type":"integer","format":"int32"},"isAvailable":{"type":"boolean"}}},"PostClassPlanDTO":{"type":"object","properties":{"gradPlanId":{"type":"integer","format":"int64"},"description":{"type":"string"},"availability":{"$ref":"#/components/schemas/PostAvailabilityDTO"},"classDistribution":{"type":"string","enum":["CONCENTRATED","SPARSE"]},"burdenCapacity":{"type":"string","enum":["LOW","MEDIUM","HIGH"]}}},"GetAvailabilityDTO":{"type":"object","properties":{"days":{"type":"array","items":{"$ref":"#/components/schemas/GetAvailabilityDayDTO"}}}},"GetAvailabilityDayDTO":{"type":"object","properties":{"day":{"type":"string"},"hours":{"type":"array","items":{"$ref":"#/components/schemas/GetAvailabilityHourDTO"}}}},"GetAvailabilityHourDTO":{"type":"object","properties":{"hourOfTheDay":{"type":"integer","format":"int32"},"isAvailable":{"type":"boolean"}}},"GetClassDetailDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"section":{"type":"string"},"instructor":{"type":"string"},"room":{"type":"string"},"method":{"type":"string"},"schedule":{"type":"array","items":{"$ref":"#/components/schemas/GetClassScheduleDTO"}}}},"GetClassPlanDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"gradPlan":{"$ref":"#/components/schemas/GetGradPlanDTO"},"description":{"type":"string"},"classes":{"type":"array","items":{"$ref":"#/components/schemas/GetCourseDTO"}},"availability":{"$ref":"#/components/schemas/GetAvailabilityDTO"},"classDistribution":{"type":"string","enum":["CONCENTRATED","SPARSE"]},"burdenCapacity":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"GetClassScheduleDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"day":{"type":"string"},"startTime":{"type":"string"},"endTime":{"type":"string"}}},"GetCourseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"subject":{"type":"string"},"number":{"type":"string"},"title":{"type":"string"},"classDetail":{"$ref":"#/components/schemas/GetClassDetailDTO"}}}}}}